"Updated UserController, ProductController, and ProductGroup model with new fields and relationships; added validation rules and error handling; modified views and routes to accommodate changes."

+++ database/migrations/2024_02_09_121556_create_product_groups_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -16,7 +15,7 @@ return new class extends Migration
             $table->string('group_name')->unique();
             $table->string('description')->nullable();
             $table->integer('profit')->nullable();
-            $table->integer('amount')->nullable();
+            $table->decimal('amount', 10, 2)->nullable();
             $table->integer('hs_de');
             $table->integer('hs_cn');
             $table->timestamps();
